+++ b2/data_source_b2_bucket.go
@@ -104,14 +104,14 @@ func dataSourceB2BucketRead(ctx context.Context, d *schema.ResourceData, meta in
 		"bucket_name": d.Get("bucket_name").(string),
 	}
 
-	output, err := client.apply(name, op, input)
+	output, err := client.apply(ctx, name, op, input)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 
 	d.SetId(output["bucket_id"].(string))
 
-	err = client.populate(name, op, output, d)
+	err = client.populate(ctx, name, op, output, d)
 	if err != nil {
 		return diag.FromErr(err)
 	}